Create a subpackage for struct PrometheusStatsCollector (#264)
Moved PrometheusStatsCollector to a new subpackage 'prometheus'.

Renamed it to StatsCollector, as the package name is descriptive enough.
